# core-common-kycb-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-kycb-mgmt
  description: KYC/B & AML Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/kycb
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/identity/parties/{partyId}/kyc/{verificationId}/edd:
    get:
      tags:
        - Enhanced Due Diligence
      summary: List EDD processes
      description: Retrieves all enhanced due diligence processes for the specified KYC verification with filtering capabilities
      operationId: listEddProcesses
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the KYC verification
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.kycVerificationId
          in: query
          description: Exact filter for kyc verification id
          required: false
          schema:
            type: string
        - name: filters.eddReason
          in: query
          description: Exact filter for edd reason
          required: false
          schema:
            type: string
        - name: filters.eddStatus
          in: query
          description: Exact filter for edd status
          required: false
          schema:
            type: string
        - name: filters.eddDescription
          in: query
          description: Exact filter for edd description
          required: false
          schema:
            type: string
        - name: filters.approvingAuthority
          in: query
          description: Exact filter for approving authority
          required: false
          schema:
            type: string
        - name: filters.approvalDate
          in: query
          description: Exact filter for approval date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[approvalDate].from
          in: query
          description: Filter approval date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[approvalDate].to
          in: query
          description: Filter approval date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.eddNotes
          in: query
          description: Exact filter for edd notes
          required: false
          schema:
            type: string
        - name: filters.internalCommitteeApproval
          in: query
          description: Exact filter for internal committee approval
          required: false
          schema:
            type: boolean
        - name: filters.committeeApprovalDate
          in: query
          description: Exact filter for committee approval date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[committeeApprovalDate].from
          in: query
          description: Filter committee approval date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[committeeApprovalDate].to
          in: query
          description: Filter committee approval date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.completionDate
          in: query
          description: Exact filter for completion date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[completionDate].from
          in: query
          description: Filter completion date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[completionDate].to
          in: query
          description: Filter completion date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.completedBy
          in: query
          description: Exact filter for completed by
          required: false
          schema:
            type: string
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved EDD processes
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    post:
      tags:
        - Enhanced Due Diligence
      summary: Start new EDD process
      description: Creates a new enhanced due diligence process for the specified KYC verification
      operationId: startEddProcess
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the KYC verification
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
        required: true
      responses:
        '201':
          description: Successfully created EDD process
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
...
  /api/v1/identity/parties/{partyId}/kyc/{verificationId}/edd/{eddId}:
    get:
      tags:
        - Enhanced Due Diligence
      summary: Get specific EDD process
      description: Retrieves a specific enhanced due diligence process by its ID
      operationId: getEddProcess
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the KYC verification
          required: true
          schema:
            type: string
            format: uuid
        - name: eddId
          in: path
          description: ID of the EDD process
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved EDD process
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
        '404':
          description: EDD process not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
    delete:
      tags:
        - Enhanced Due Diligence
      summary: Delete EDD process
      description: Deletes an enhanced due diligence process
      operationId: deleteEddProcess
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the KYC verification
          required: true
          schema:
            type: string
            format: uuid
        - name: eddId
          in: path
          description: ID of the EDD process
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted EDD process
        '404':
          description: EDD process not found
    patch:
      tags:
        - Enhanced Due Diligence
      summary: Update EDD process
      description: Updates an existing enhanced due diligence process
      operationId: updateEddProcess
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the KYC verification
          required: true
          schema:
            type: string
            format: uuid
        - name: eddId
          in: path
          description: ID of the EDD process
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
        required: true
      responses:
        '200':
          description: Successfully updated EDD process
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
        '404':
          description: EDD process not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnhancedDueDiligenceDTO'
...
components:
  schemas:
...
    EnhancedDueDiligenceDTO:
      required:
        - approvalDate
        - approvingAuthority
        - eddDescription
        - eddReason
        - eddStatus
        - internalCommitteeApproval
        - kycVerificationId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        eddId:
          type: string
          format: uuid
          readOnly: true
        kycVerificationId:
          type: string
          format: uuid
        eddReason:
          maxLength: 100
          minLength: 0
          type: string
        eddStatus:
          maxLength: 50
          minLength: 0
          type: string
        eddDescription:
          maxLength: 1000
          minLength: 0
          type: string
        approvingAuthority:
          maxLength: 200
          minLength: 0
          type: string
        approvalDate:
          type: string
          format: date-time
        eddNotes:
          maxLength: 2000
          minLength: 0
          type: string
        internalCommitteeApproval:
          type: boolean
        committeeApprovalDate:
          type: string
          format: date-time
        completionDate:
          type: string
          format: date-time
        completedBy:
          maxLength: 100
          minLength: 0
          type: string
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.